Update Core-v3.0.ps1
I am using Zergpool and mining to LTC instead of BTC as LTC is the other payment coin guaranteed by the pool. I made a few changes shown here to show the proper conversion prices. I was going to make more changes but I don't fully understand the implications of changing this across the board. My main concern is that it is going to screw up the plus brain data due to the difference in expected BTC versus the actual LTC I receive. I know it messes with the expected payout date which doesn't bother me but I hope the plus data is still effective.