Kill miners by path and exe name
In rare cases, particularly on systems with low end CPUs that are laggy
when mining, the process handle for a miner can be incorrect. This
causes a miner to appear to have exited or failed even though it is
still running.

When the miner is restarted, or a new miner started, the old one will
still be mining, and worse listening on that port, so the new miner's
stats will get messed up.

Adds a couple extra lines to find such miners by their path and kill
them even if the process handle is incorrect.