v2.9.0
This release introduce cryptonight_r required for the Monero for on 2019/09/03. Select the currency monero to switch automatically the algorithm from cryptonight_v8 to cryptonight_r with the Monero blockversion 10.

Please note that this version is a bit rough around the edges due to the emergency nature of Monero fork. CPU mining in particular will be very slow. We plan to release a more polished version around the fork date.